# swagger-docs.json
{"openapi":"3.0.0","paths":{"/user":{"get":{"operationId":"UserController_getUser","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["user"]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["user"]}},"/user/create":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["user"]}},"/user/update/{id}":{"put":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["user"]}},"/user/delete/{id}":{"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["user"]}},"/auth":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"]}},"/reports":{"get":{"operationId":"ReportsController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["reports"]}},"/reports/{id}":{"delete":{"operationId":"ReportsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["reports"]}},"/post/create":{"post":{"operationId":"PostController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePostDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["post"]}},"/post":{"get":{"operationId":"PostController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["post"]}},"/post/{id}":{"get":{"operationId":"PostController_findAllByUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["post"]}},"/post/update/{id}":{"put":{"operationId":"PostController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePostDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["post"]}},"/post/delete/{id}":{"delete":{"operationId":"PostController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["post"]}},"/reports/administration":{"post":{"operationId":"AdministrationController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAdministrationReportDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["reports"]}},"/reports/administration/{id}":{"get":{"operationId":"AdministrationController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["reports"]},"patch":{"operationId":"AdministrationController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAdministrationDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["reports"]},"delete":{"operationId":"AdministrationController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiCommonResponse"}}}}},"security":[{"bearer":[]}],"tags":["reports"]}}},"info":{"title":"Nest API","description":"The Nest API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ApiCommonResponse":{"type":"object","properties":{"statusCode":{"type":"number"},"data":{"type":"object"},"error":{"type":"string"},"message":{"type":"string"}},"required":["statusCode"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"role":{"enum":["gerente","admin","user"],"type":"string"},"password":{"type":"string"}},"required":["name","email","role","password"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}},"required":["name","email"]},"AuthLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreatePostDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}},"required":["name","description"]},"UpdatePostDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"}},"required":["id","name","description"]},"CreateAdministrationReportDto":{"type":"object","properties":{"reportType":{"type":"object"},"farm":{"type":"string"},"manager":{"type":"string"},"correctUseOfEPI":{"type":"number"}},"required":["reportType","farm","manager","correctUseOfEPI"]},"UpdateAdministrationDto":{"type":"object","properties":{"reportType":{"type":"object"},"farm":{"type":"string"},"manager":{"type":"string"},"correctUseOfEPI":{"type":"number"}},"required":["reportType","farm","manager","correctUseOfEPI"]}}}}